Fix AfricasTalking outboxId JSON type mismatch dropping all messages
PHP publishes outboxId as a quoted string on AFRICAS_TALKING_SMS_QUEUE,
but domain.Message.OutboxID was typed int64. Every message failed
Consumer.parseMessages as a result and was Nacked without requeue —
silently dropped, with no error trace tying it back to AfricasTalking
specifically (the failure log lacked queue context). SDP/SMPP never
populate outboxId, so they were never affected; this is 100% of
AfricasTalking traffic.

Adds FlexibleInt64, which accepts a JSON number or a quoted numeric
string, and uses it for OutboxID. Includes regression tests
reproducing the exact failing PHP payload.

+++ internal/sms/adapters/httpapi/africas_talking_reporter.go
@@ -67,7 +67,7 @@ func (r *AfricasTalkingReporter) Report(ctx context.Context, result *domain.Send
 	}
 
 	payload := sendResultPayload{
-		OutboxID:    result.Message.OutboxID,
+		OutboxID:    int64(result.Message.OutboxID),
 		Success:     success,
 		Status:      status,
 		NetworkCode: result.NetworkCode,
